requirements.txt
aiosqlite==0.17.0
anyio==3.3.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.2.0
autopep8==1.5.7
boto3==1.20.38
botocore==1.23.38
cffi==1.14.6
databases==0.5.4
fastapi==0.72.0
greenlet==1.1.1
idna==3.3
iniconfig==1.1.1
jmespath==0.10.0
orjson==3.6.5
ormar==0.10.24
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.23
starlette==0.17.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
watchtower==2.1.1